only update content if it has changed

+++ scripts/generate-pkg-readmes.mjs
@@ -8,7 +8,9 @@ Object.entries(pkgDetails).forEach(([dir, pkgJson]) => {
     const readmePath = path.join(dir, 'README.md');
     const esmShHref = `https://esm.sh/jsr/@web-components/${pkgJson.name}`;
     const readmeContent = `# ${pkgJson.name}\n\n**version:** ${pkgJson.version}\n\n> ${pkgJson.description}\n\nUse in a browser with [${esmShHref}](${esmShHref})\n\nMade by [jackcarey](https://jackcarey.co.uk).`;
-    fs.writeFileSync(readmePath, readmeContent);
+    if (!fs.existsSync(readmePath) || fs.readFileSync(readmePath, 'utf8') !== readmeContent) {
+        fs.writeFileSync(readmePath, readmeContent);
+    }
 });
 
 console.log(count + ' package details saved to README files');
